a little more robust
Some common errors are handled now. Because of the magnitude of the problem i.e making every whois record work, there is a lot of code that is "just written".  It will take some time to clean all of this up, but it will be worthwhile. One python whois to rule them all.